# 7agz.Core/DTOs/PlayerRateDTO.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace _7agz.Core.DTOs
{
    public class PlayerRateDTO
    {
        public decimal pace { get; set; } = 67.7008989m;
        public decimal defending { get; set; } = 51.55350326m;
        public decimal passing { get; set; } = 57.23377663m;
        public decimal shooting { get; set; } = 52.2983007m;
        public decimal dribbling { get; set; } = 62.53158478m;
        

    }
}